Create README.md
INTRODUCTION TO DATA STRUCTURES ALGORITHM
presentation to Information structure and calculations Presentation TO Information STRUCTURE AND Calculations : ** Introduction to Information Structures and Calculations ** Data Structure could be a way of collecting and organising data in such a way that we are able perform operations on these information in an compelling way. Information Structures is approximately rendering information components in terms of a few relationship, for way better organization and capacity. For case, we have a few information which has, player's title "Virat" and age 26. Here "Virat" is of String information sort and 26 is of numbers information type. We can organize this information as a record like Player record, which is able have both player's title and age in it. Presently we will collect and store player's records in a record or database as a information structure. For case: "Dhoni" 30, "Gambhir" 31, "Sehwag" 33 In case you're mindful of Protest Situated programming concepts, at that point a course too does the same thing, it collects distinctive sort of information beneath one single substance. The as it were contrast being, information structures gives for methods to get to and control information efficiently. In straightforward dialect, Information Structures are structures modified to store requested information, so that different operations can be performed on it effectively. It speaks to the information of information to be organized in memory. It ought to be planned and actualized in such a way that it decreases the complexity and increments the efficiency. Basic sorts of Information Structures As we have examined over, anything that can store information can be called as a information structure, consequently Numbers, Drift, Boolean, Char etc, all are information structures. They are known as Primitive Information Structures. Then we moreover have a few complex Information Structures, which are utilized to store expansive and associated information. A few illustration of Theoretical Information Structure are : Linked List Tree Graph Stack, Queue etc. All these data structures allow us to perform different operations on data. We select these data structures based on which type of operation is required. We will look into these data structures in more details in our later lessons.

Introduction to Data Structures The data structures can also be classified on the basis of the following characteristics:

Characterstic Description Linear In Linear data structures,the data items are arranged in a linear sequence. Example: Array Non-Linear In Non-Linear data structures,the data items are not in sequence. Example: Tree, Graph Homogeneous In homogeneous data structures,all the elements are of same type. Example: Array Non-Homogeneous In Non-Homogeneous data structure, the elements may or may not be of the same type. Example: Structures Static Static data structures are those whose sizes and structures associated memory locations are fixed, at compile time. Example: Cluster Energetic Energetic structures are those which grows or shrivels depending upon the program require and its execution. Moreover, their related memory areas changes. Case: Connected List made utilizing pointers What is an Calculation ? An calculation could be a limited set of instructions or rationale, composed in arrange, to achieve a certain predefined errand. Calculation isn't the total code or program, it is fair the center logic(solution) of a issue, which can be communicated either as an casual tall level portrayal as pseudocode or employing a flowchart. Every Calculation must fulfill the taking after properties: Input- There ought to be or more inputs provided remotely to the calculation. Yield- There ought to be atleast 1 yield gotten. Definiteness- Each step of the calculation ought to be clear and well characterized. Limit- The calculation ought to have limited number of steps. Rightness- Each step of the calculation must create a rectify yield. An calculation is said to be proficient and quick, in the event that it takes less time to execute and expends less memory space. The execution of an calculation is measured on the premise of taking after properties : _ Time Complexity Space Complexity Space Complexity_ Its the sum of memory space required by the calculation, amid the course of its execution. Space complexity must be taken truly for multi-user frameworks and in circumstances where constrained memory is available. An calculation for the most part requires space for taking after components : Instruction Space: Its the space required to store the executable adaptation of the program. This space is settled, but shifts depending upon the number of lines of code within the program. Information Space: Its the space required to store all the constants and variables(including transitory factors) esteem. Environment Space: Its the space required to store the environment data required to continue the suspended work. To memorize almost Space Complexity in detail, hop to the Space Complexity tutorial. Time Complexity Time Complexity may be a way to speak to the sum of time required by the program to run till its completion. It's by and large a great hone to undertake to keep the time required minimum, so that our calculation completes it's execution within the least time conceivable. We are going ponder approximately Time Complexity in subtle elements in afterward areas.